--- components/MuiButton/index.ts
@@ -1,1 +1,0 @@
-export * from './muibutton.component';
