+++ components/index.ts
@@ -1,3 +1,3 @@
+export * from './Button';
 export * from './LogoHeader';
 export * from './Modal';
-export * from './MuiButton';
